requirements.txt
streamlit==1.25.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.12.0
scikit-learn==1.2.2
pandas==2.1.2
numpy==1.26.0
pyyaml==6.0
tqdm==4.66.1
python-dotenv==1.0.0
